itest: moving routing related tests into one file
